Updated Settings pages. combined seperate viewmodels into one viewmodel for simplisty sake, hopefully this will not make things super complex in the future... for now it makes sense though. Added Scrollview to unit selection page.
