+++ OilfieldCalc3/Resx/AppResources.resx
@@ -207,6 +207,18 @@
   <data name="String_Capacty" xml:space="preserve">
     <value>Capacity</value>
   </data>
+  <data name="String_ClearData" xml:space="preserve">
+    <value>Clear Data</value>
+  </data>
+  <data name="String_ClearDrillstring_Checkbox" xml:space="preserve">
+    <value>Clear Drillstring Database</value>
+  </data>
+  <data name="String_ClearSettings_Checkbox" xml:space="preserve">
+    <value>Clear User Settings</value>
+  </data>
+  <data name="String_ClearWellbore_Checkbox" xml:space="preserve">
+    <value>Clear Wellbore Database</value>
+  </data>
   <data name="String_ColorOptions" xml:space="preserve">
     <value>Color Options</value>
     <comment>String = Color Options</comment>
