+++ OilfieldCalc3/Views/SettingsAdvancedPage.xaml
@@ -1,12 +1,13 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:syncfusion="clr-namespace:Syncfusion.XForms.Buttons;assembly=Syncfusion.Buttons.XForms"
              xmlns:prism="http://prismlibrary.com"
              xmlns:resources="clr-namespace:OilfieldCalc3.Resx"
              xmlns:viewTemplate="clr-namespace:OilfieldCalc3.Views"
              prism:ViewModelLocator.AutowireViewModel="True"
              x:Class="OilfieldCalc3.Views.SettingsAdvancedPage"
-             Title="{Binding Title}">
+             Title="{Binding SettingsAdvancedPageTitle}">
     <viewTemplate:ViewPageTemplate>
         <viewTemplate:ViewPageTemplate.Header>
             <StackLayout>
@@ -22,8 +23,9 @@
                                       TextColor="{AppThemeBinding Dark={StaticResource DarkPrimaryTextColor}, Light={StaticResource LightPrimaryTextColor}}">
                             <ViewCell>
                                 <StackLayout HorizontalOptions="StartAndExpand">
-                                    <Button Text="Clear Database" 
-                                            Command="{Binding WipeDatabaseCommand}"/>
+                                    <syncfusion:SfButton Text="Clear Database" 
+                                                         Command="{Binding WipeDatabaseCommand}"
+                                                         Style="{StaticResource sfButton}"/>
                                 </StackLayout>
                             </ViewCell>
                         </TableSection>
