+++ OilfieldCalc3/Views/SettingsUnitsPage.xaml
@@ -7,7 +7,7 @@
              xmlns:viewTemplate="clr-namespace:OilfieldCalc3.Views"
              prism:ViewModelLocator.AutowireViewModel="True"
              x:Class="OilfieldCalc3.Views.SettingsUnitsPage"
-             Title="{Binding Title}">
+             Title="{Binding SettingsUnitsPageTitle}">
     <viewTemplate:ViewPageTemplate>
         <viewTemplate:ViewPageTemplate.Header>
             <StackLayout>
@@ -16,66 +16,68 @@
         </viewTemplate:ViewPageTemplate.Header>
 
         <viewTemplate:ViewPageTemplate.Body>
-            <StackLayout>
-                <TableView Intent="Settings">
-                    <TableRoot>
-                        <TableSection Title="{x:Static resources:AppResources.String_LongLengthUnit}"
+            <ScrollView>
+                <StackLayout>
+                    <TableView Intent="Settings">
+                        <TableRoot>
+                            <TableSection Title="{x:Static resources:AppResources.String_LongLengthUnit}"
                                       TextColor="{AppThemeBinding Dark={StaticResource DarkPrimaryTextColor}, Light={StaticResource LightPrimaryTextColor}}">
-                            <ViewCell>
-                                <StackLayout Orientation="Horizontal" Padding="20,0,0,0">
-                                    <Picker ItemsSource="{Binding LongLengthUnits}"
-                                            HorizontalOptions="StartAndExpand"
+                                <ViewCell>
+                                    <StackLayout Orientation="Horizontal" Padding="20,0,0,0" HorizontalOptions="FillAndExpand">
+                                        <Picker ItemsSource="{Binding LongLengthUnits}"
+                                            HorizontalOptions="FillAndExpand"
                                             SelectedItem="{Binding SelectedLongLengthUnit, Mode=TwoWay}"/>
 
-                                </StackLayout>
-                            </ViewCell>
-                        </TableSection>
-                        <TableSection Title="{x:Static resources:AppResources.String_ShortLengthUnit}"
+                                    </StackLayout>
+                                </ViewCell>
+                            </TableSection>
+                            <TableSection Title="{x:Static resources:AppResources.String_ShortLengthUnit}"
                                       TextColor="{AppThemeBinding Dark={StaticResource DarkPrimaryTextColor}, Light={StaticResource LightPrimaryTextColor}}">
-                            <ViewCell>
-                                <StackLayout Orientation="Horizontal" Padding="20,0,0,0">
-                                    <Picker ItemsSource="{Binding ShortLengthUnits}"
-                                            HorizontalOptions="StartAndExpand"
+                                <ViewCell>
+                                    <StackLayout Orientation="Horizontal" Padding="20,0,0,0" HorizontalOptions="FillAndExpand">
+                                        <Picker ItemsSource="{Binding ShortLengthUnits}"
+                                            HorizontalOptions="FillAndExpand"
                                             SelectedItem="{Binding SelectedShortLengthUnit, Mode=TwoWay}"/>
 
-                                </StackLayout>
-                            </ViewCell>
-                        </TableSection>
-                        <TableSection Title="{x:Static resources:AppResources.String_VolumeUnit}"
+                                    </StackLayout>
+                                </ViewCell>
+                            </TableSection>
+                            <TableSection Title="{x:Static resources:AppResources.String_VolumeUnit}"
                                       TextColor="{AppThemeBinding Dark={StaticResource DarkPrimaryTextColor}, Light={StaticResource LightPrimaryTextColor}}">
-                            <ViewCell>
-                                <StackLayout Orientation="Horizontal" Padding="20,0,0,0">
-                                    <Picker ItemsSource="{Binding VolumeUnits}"
-                                            HorizontalOptions="StartAndExpand"
+                                <ViewCell>
+                                    <StackLayout Orientation="Horizontal" Padding="20,0,0,0" HorizontalOptions="FillAndExpand">
+                                        <Picker ItemsSource="{Binding VolumeUnits}"
+                                            HorizontalOptions="FillAndExpand"
                                             SelectedItem="{Binding SelectedVolumeUnit, Mode=TwoWay}"/>
 
-                                </StackLayout>
-                            </ViewCell>
-                        </TableSection>
-                        <TableSection Title="{x:Static resources:AppResources.String_CapacityUnit}"
+                                    </StackLayout>
+                                </ViewCell>
+                            </TableSection>
+                            <TableSection Title="{x:Static resources:AppResources.String_CapacityUnit}"
                                       TextColor="{AppThemeBinding Dark={StaticResource DarkPrimaryTextColor}, Light={StaticResource LightPrimaryTextColor}}">
-                            <ViewCell>
-                                <StackLayout Orientation="Horizontal" Padding="20,0,0,0">
-                                    <Picker ItemsSource="{Binding CapacityUnits}"
-                                            HorizontalOptions="StartAndExpand"
+                                <ViewCell>
+                                    <StackLayout Orientation="Horizontal" Padding="20,0,0,0" HorizontalOptions="FillAndExpand">
+                                        <Picker ItemsSource="{Binding CapacityUnits}"
+                                            HorizontalOptions="FillAndExpand"
                                             SelectedItem="{Binding SelectedCapacityUnit, Mode=TwoWay}"/>
 
-                                </StackLayout>
-                            </ViewCell>
-                        </TableSection>
-                        <TableSection Title="{x:Static resources:AppResources.String_MassUnit}"
+                                    </StackLayout>
+                                </ViewCell>
+                            </TableSection>
+                            <TableSection Title="{x:Static resources:AppResources.String_MassUnit}"
                                       TextColor="{AppThemeBinding Dark={StaticResource DarkPrimaryTextColor}, Light={StaticResource LightPrimaryTextColor}}">
-                            <ViewCell>
-                                <StackLayout Orientation="Horizontal" Padding="20,0,0,0">
-                                    <Picker ItemsSource="{Binding MassUnits}"
-                                            HorizontalOptions="StartAndExpand"
+                                <ViewCell>
+                                    <StackLayout Orientation="Horizontal" Padding="20,0,0,0" HorizontalOptions="FillAndExpand">
+                                        <Picker ItemsSource="{Binding MassUnits}"
+                                            HorizontalOptions="FillAndExpand"
                                             SelectedItem="{Binding SelectedMassUnit, Mode=TwoWay}"/>
-                                </StackLayout>
-                            </ViewCell>
-                        </TableSection>
-                    </TableRoot>
-                </TableView>
-            </StackLayout>
+                                    </StackLayout>
+                                </ViewCell>
+                            </TableSection>
+                        </TableRoot>
+                    </TableView>
+                </StackLayout>
+            </ScrollView>
         </viewTemplate:ViewPageTemplate.Body>
 
         <viewTemplate:ViewPageTemplate.Footer>
